Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/content-security-policy/reporting/report-original-url-on-mixed-content-frame.https.sub-expected.txt


FAIL Violation report status OK. assert_equals: No such report. expected "" but got "false"
